RP2A/requirements.txt
fastmcp
PyJWT
rank-bm25
nltk
python-dotenv
starlette
uvicorn